--- src/test/java/roomescape/reservation/service/PaymentServiceTest.java
@@ -1,12 +1,0 @@
-package roomescape.reservation.service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PaymentServiceTest {
-
-    @Test
-    void payment() {
-    }
-}
